+++ packages/web-server/src/application/repositories/reports-repository.ts
@@ -1,4 +1,4 @@
-import { Report } from '../../enterprise/entities';
+import { Report } from '../../domain/models';
 
 export interface ReportsRepository {
   findByDiagnosisId(id: string): Promise<Report[]>;
