+++ packages/web-server/src/domain/models/rule.ts
