+++ packages/web-server/src/frameworks/database/repositories/diagnosis-repository-impl.ts
@@ -2,7 +2,7 @@ import { inject, injectable } from 'inversify';
 import { Connection } from 'typeorm';
 
 import { DiagnosisRepository } from '../../../application/repositories/diagnosis-repository';
-import { Diagnosis, Report } from '../../../enterprise/entities';
+import { Diagnosis, Report } from '../../../domain/models';
 import { TYPES } from '../../../types';
 import { DiagnosisORM } from '../entities/diagnosis';
 import { ReportORM } from '../entities/report';
