+++ packages/web-server/src/frameworks/database/repositories/reports-repository-in-memory-impl.ts
@@ -1,7 +1,7 @@
 import { injectable, unmanaged } from 'inversify';
 
 import { ReportsRepository } from '../../../application/repositories/reports-repository';
-import { Report } from '../../../enterprise/entities';
+import { Report } from '../../../domain/models';
 
 @injectable()
 export class ReportsRepositoryInMemoryImpl implements ReportsRepository {
